+++ src/main/java/faang/school/postservice/client/ProjectServiceClient.java
@@ -12,14 +12,11 @@ import java.util.List;
 @FeignClient(name = "project-service", url = "${project-service.host}:${project-service.port}", path = "/api/v1/projects")
 public interface ProjectServiceClient {
 
-    @GetMapping("/api/v1/project/{projectId}")
+    @GetMapping("/{projectId}")
     ProjectDto getProject(@PathVariable("projectId") long projectId);
 
-    @PostMapping("/api/v1/projects")
+    @PostMapping
     List<ProjectDto> getProjectsByIds(@RequestBody List<Long> ids);
-
-    @PostMapping("/api/v1/projects/exists/{projectId}")
-    void existsProjectById(@PathVariable("projectId") long projectId);
 
     @GetMapping
     List<ProjectDto> getAll();
