+++ src/main/java/faang/school/postservice/controller/PostController.java
@@ -6,11 +6,7 @@ import faang.school.postservice.service.PostService;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.DeleteMapping;
-import faang.school.postservice.dto.post.PostDto;
-import faang.school.postservice.service.PostService;
-import jakarta.validation.Valid;
 import jakarta.validation.constraints.Size;
-import lombok.RequiredArgsConstructor;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -19,9 +15,7 @@ import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestPart;
-import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.multipart.MultipartFile;
 
 import java.util.List;
@@ -43,11 +37,6 @@ public class PostController {
     @PostMapping("/publish/{postId}")
     public void publishPost(@PathVariable long postId) {
         postService.publishPost(postId, userContext.getUserId());
-    }
-
-    @PutMapping("/{postId}")
-    public void updatePost(@PathVariable long postId, @RequestBody PostDto dto) {
-        postService.updatePost(postId, userContext.getUserId(), dto);
     }
 
     @DeleteMapping("/{postId}")
