+++ src/main/java/faang/school/postservice/dto/PostDto.java
@@ -2,17 +2,21 @@ package faang.school.postservice.dto;
 
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
+import jakarta.validation.constraints.Size;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
 import lombok.Data;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
 import java.time.LocalDateTime;
+import java.util.List;
 
 @Data
 @NoArgsConstructor
-@Getter
-@Setter
+@AllArgsConstructor
+@Builder
 public class PostDto {
     private Long id;
     @NotNull
@@ -25,5 +29,6 @@ public class PostDto {
     private LocalDateTime updatedAt;
     private boolean published;
     private boolean deleted;
-
+    @Size(max = 10)
+    private List<Long> resourceIds;
 }
