+++ src/test/java/faang/school/postservice/validator/PostValidatorTest.java
@@ -2,8 +2,9 @@ package faang.school.postservice.validator;
 
 import faang.school.postservice.client.ProjectServiceClient;
 import faang.school.postservice.client.UserServiceClient;
+import faang.school.postservice.config.context.UserContext;
 import faang.school.postservice.dto.PostDto;
-import faang.school.postservice.dto.post.PostDto;
+import faang.school.postservice.dto.ProjectDto;
 import faang.school.postservice.exception.DataValidationException;
 import faang.school.postservice.model.Post;
 import faang.school.postservice.repository.PostRepository;
@@ -35,6 +36,9 @@ public class PostValidatorTest {
 
     @Mock
     private PostRepository postRepository;
+
+    @Mock
+    UserContext userContext;
 
     @InjectMocks
     private PostValidator postValidator;
